+++ src/js/player.js
@@ -1924,6 +1924,19 @@ class Player extends Component {
   }
 
   /**
+   * Reset the player. Loads the first tech in the techOrder,
+   * and calls `reset` on the tech`.
+   *
+   * @return {Player} Returns the player
+   * @method reset
+   */
+  reset() {
+    this.loadTech_(toTitleCase(this.options_.techOrder[0]), null);
+    this.techCall_('reset');
+    return this;
+  }
+
+  /**
    * Returns the fully qualified URL of the current source value e.g. http://mysite.com/video.mp4
    * Can be used in conjuction with `currentType` to assist in rebuilding the current source object.
    *